% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_plastics.R
\name{clean_plastics}
\alias{clean_plastics}
\title{Cleans and recodes raw plastics data}
\usage{
clean_plastics(data)
}
\arguments{
\item{data}{A data frame from \code{\link[=load_data]{load_data()}}.}
}
\value{
A data.table with corrected country names and added
\verb{_prop} columns for each plastic type.
}
\description{
Standardizes country names, removes "bad" rows, and adds plastic proportion
columns. Uses data.table for better in-place mutation.
}
